requirements.txt
langchain
langchain-community
pypdf
sentence-transformers 
pinecone
dotenv
fastapi
uvicorn
requests
faiss-cpu